requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
language-tool-python
spacy
reportlab
openai
pytest
httpx
pytest-mock
